fix: move types
