+++ src/common/language.ts
@@ -3,5 +3,3 @@ export enum Language {
   UA = 'ua',
   EN = 'en'
 }
-
-export type TranslatedField = Record<Language, string>
